+++ src/libraries/gaming/PopulationFlow.h
@@ -48,7 +48,7 @@ public:
     virtual unsigned long GetId() const;
     virtual unsigned int GetLivingHumans() const;
     virtual unsigned int GetDeadHumans() const;
-    virtual unsigned int GetDensity() const;
+    virtual float GetDensity() const;
     virtual QString      GetAttitude() const;
     virtual float GetHeight() const;
